+++ src/causal_rag_audit/__init__.py
@@ -7,14 +7,27 @@ from .models import (
     AuditDataset,
     AuditReport,
     Document,
+    MeanMetricResult,
     TargetRequest,
     TargetResponse,
 )
 from .runner import AuditRunner, ResponseFormatError
+from .scoring import (
+    PAPER_PROFILE,
+    SCORING_PROFILES,
+    STRICT_PROFILE,
+    AbstentionJudge,
+    AnswerJudge,
+)
 from .validation import DatasetValidationError, dataset_from_mapping, load_dataset
 from .version import __version__
 
 __all__ = [
+    "PAPER_PROFILE",
+    "SCORING_PROFILES",
+    "STRICT_PROFILE",
+    "AbstentionJudge",
+    "AnswerJudge",
     "AuditCase",
     "AuditDataset",
     "AuditReport",
@@ -23,6 +36,7 @@ __all__ = [
     "Document",
     "HttpTarget",
     "KeywordTarget",
+    "MeanMetricResult",
     "ResponseFormatError",
     "TargetRequest",
     "TargetResponse",